requirements.txt
python-telegram-bot==20.7
meshtastic==2.2.16
pydantic==1.10.15
python-dotenv==1.0.1
click==8.1.7
rich==13.7.1
black==24.3.0
pylint==3.2.6
pre-commit==3.7.1
pytest==8.3.2
